fixed compare functions not using big enough return type

+++ src/ft_ls.h
@@ -6,7 +6,7 @@
 /*   By: mcarter <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/18 13:02:44 by mcarter           #+#    #+#             */
-/*   Updated: 2019/08/23 09:38:31 by mcarter          ###   ########.fr       */
+/*   Updated: 2019/08/23 12:35:07 by mcarter          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -115,9 +115,9 @@ t_elem			*get_folder(char *path, t_args args);
 MAXUNBR			get_folder_count(char *path, unsigned char all);
 int				filter(unsigned char all, char *name);
 void			sort_elem_array(t_elem *arr, t_args args);
-int				cmp_name(t_elem elem1, t_elem elem2);
-int				cmp_modified(t_elem elem1, t_elem elem2);
-int				cmp_access(t_elem elem1, t_elem elem2);
+long			cmp_name(t_elem elem1, t_elem elem2);
+long			cmp_modified(t_elem elem1, t_elem elem2);
+long			cmp_access(t_elem elem1, t_elem elem2);
 void			show_folder_subfolders(char *path, t_args args, t_elem *items);
 void			free_items(t_elem **items);
